Using Plato to generate the analysis data
