+++ packages/types/src/index.ts
@@ -2,11 +2,11 @@ import { PlatformPath } from "path";
 import { type ArgumentsCamelCase } from "yargs";
 
 export interface AnalyzeCommandArgs extends ArgumentsCamelCase {
+  analyzer: "plato";
   files: string[];
-  outputDir: string;
-  recursive: boolean;
-  title: string;
   exclude: string[];
+  outputDir: string;
+  title: string;
 }
 
 export interface ReportCommandArgs extends ArgumentsCamelCase {
